// src/helper/types.ts
export interface HomeSliceProps {
  bannerList: BannerListProps[];
  hotCoins: CoinDataProps[];
  newListedCoins: CoinDataProps[];
  coinPairs: CoinDataProps[];
  hotPairsChart?: Record<string, number[]>;
  currency: string;
  coinBalance: Object;
  coinData: CoinDataProps[];
  spotSelectedPair?: CoinDataProps | null;
  futuresSelectedPair?: any | null;
  favorites: CoinDataProps[];
  notificationList: NotificationProps[];
  referCode?: number;
  referCount: number;
  spotOpenOrders: [];
  socketLoading: boolean;
  crossAccount?: any;
  crossBorrowable?: any;
  favoriteArray: string[];
  favoriteArrayLoaded: boolean;
  openOrders: any[] | null;
  pastOrders: any[];
  fiveDaySymbolData: HistoricDataProps[];
  oneMonthSymbolData: HistoricDataProps[];
  ThreeMonthSymbolData: HistoricDataProps[];
  oneYearSymbolData: HistoricDataProps[];
  fiveYearSymbolData: HistoricDataProps[];
  socket?: socketIoClient.Socket<any>;
  feeDetails?: any;
  buyOrders?: OrdersProps[];
  sellOrders?: OrdersProps[];
  orderData?: any;
  futureOrders: FutureOrdersState;
  futurePositions: any[];
  random?: any;
  twoFaQrData?: any;
  conversionHistory?: any[];
  conversion?: any;
  coinList?: any;
  futuresPairs?: any[];
  qbsHistory?: [];
  stakingHome?: any;
  lakedHistory?:any;
  stakingHistory?:[];
  p2p?:any;
  userEligibility: any;
  allProjectsList: any;
  checkCommitExistense: any;
  userCommitProject: any;
  pastAllProjects: any;
  userCommits: any;
  singleProject: any;
  userProjectTotalCommit: any;
  userProjectUpdateCommit: any;
  commitDetails: any;
  activityLogs: any;
  referralList: any;
  memeList: any;
  payoutHistory: any;
  treeRoot: any;
  flatInvestments: any;
  userTickets: [],
  ticketChats: [],
}

export interface FutureOrdersState {
  openOrders: any[];
  ordersHistory: any[];
  closePositions: any[];
  tradeHistory: any[];
}
// ...
export interface CoinDataProps {
  __v: number;
  _id: string;
  available: string;
  base_currency: string;
  base_currency_id: string;
  buy_price: number;
  change: number;
  createdAt: string;
  high: number;
  icon_path: string;
  low: number;
  open: number;
  quote_currency: string;
  quote_currency_id: string;
  sell_price: number;
  status: string;
  type: string;
  updatedAt: string;
  volume: number;
}
export interface CoinCardProps {
  item: CoinDataProps;
  index: number;
  currency?:any;
}

export interface BannerListProps {
  __v: number;
  _id: string;
  bannerPath?: string;
  bannerType?: string;
  createdAt: string;
  sequrence?: string;
  status: string;
  updatedAt: string;
  banner_path?: string;
  banner_sequence?: string;
  banner_type?: string;
}
// ...
export interface NotificationProps {
  __v: number;
  _id: string;
  createdAt: string;
  message: string;
  subject: string;
  updatedAt: string;
}
// ...
export interface HistoricDataProps {
  _id: string;
  base_currency_id: string;
  quote_currency_id: string;
  open: number;
  close: number;
  high: number;
  low: number;
  volume: number;
  time: number;
}

export interface GetFeeDetailProps {
  currency_id: string;
}

export interface DeleteAccountProps {
  status: string;
}

export interface DownloadTradeReportProps {
  range: string;
}
export interface OrdersProps {
  _id: string;
  user_id: string;
  order_type: string;
  base_currency_id: string;
  quote_currency_id: string;
  ask_currency: string;
  side: string;
  price: number;
  quantity: number;
  filled: number;
  remaining: number;
  maker_fee: number;
  taker_fee: number;
  status: string;
  transaction_fee: number;
  tds: number;
  createdAt: string;
  updatedAt: string;
  __v: number;
}